+++ src/module/resolvers/file.rs
@@ -227,4 +227,30 @@ impl ModuleResolver for FileModuleResolver {
 
         Ok(m)
     }
+
+    fn resolve_ast(
+        &self,
+        engine: &Engine,
+        path: &str,
+        pos: Position,
+    ) -> Result<Option<crate::AST>, Box<EvalAltResult>> {
+        // Construct the script file path
+        let mut file_path = self.base_path.clone();
+        file_path.push(path);
+        file_path.set_extension(&self.extension); // Force extension
+
+        // Load the script file and compile it
+        let mut ast = engine
+            .compile_file(file_path.clone())
+            .map_err(|err| match *err {
+                EvalAltResult::ErrorSystem(_, err) if err.is::<IoError>() => {
+                    Box::new(EvalAltResult::ErrorModuleNotFound(path.to_string(), pos))
+                }
+                _ => Box::new(EvalAltResult::ErrorInModule(path.to_string(), err, pos)),
+            })?;
+
+        ast.set_source(path);
+
+        Ok(Some(ast))
+    }
 }